// TestAssignment.AddressBook.Ui/ClientApp/src/api/api.phone.client.ts
export class PagedListDtoOfContactDto implements IPagedListDtoOfContactDto {
    list!: ContactDto[] | undefined;
    paging!: PagingDto | undefined;

    constructor(data?: IPagedListDtoOfContactDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any, _mappings?: any) {
        if (_data) {
            if (Array.isArray(_data["list"])) {
                this.list = [] as any;
                for (let item of _data["list"])
                    this.list!.push(ContactDto.fromJS(item, _mappings));
            }
            this.paging = _data["paging"] ? PagingDto.fromJS(_data["paging"], _mappings) : <any>undefined;
        }
    }

    static fromJS(data: any, _mappings?: any): PagedListDtoOfContactDto {
        data = typeof data === 'object' ? data : {};
        return createInstance<PagedListDtoOfContactDto>(data, _mappings, PagedListDtoOfContactDto);
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        if (Array.isArray(this.list)) {
            data["list"] = [];
            for (let item of this.list)
                data["list"].push(item.toJSON());
        }
        data["paging"] = this.paging ? this.paging.toJSON() : <any>undefined;
        return data; 
    }
}

export interface IPagedListDtoOfContactDto {
    list: ContactDto[] | undefined;
    paging: PagingDto | undefined;
}

export class ContactDataDto implements IContactDataDto {
    firstName!: string;
    lastName!: string;
    phoneNumber!: string;
    address!: string;

    constructor(data?: IContactDataDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any, _mappings?: any) {
        if (_data) {
            this.firstName = _data["firstName"];
            this.lastName = _data["lastName"];
            this.phoneNumber = _data["phoneNumber"];
            this.address = _data["address"];
        }
    }

    static fromJS(data: any, _mappings?: any): ContactDataDto {
        data = typeof data === 'object' ? data : {};
        return createInstance<ContactDataDto>(data, _mappings, ContactDataDto);
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["firstName"] = this.firstName;
        data["lastName"] = this.lastName;
        data["phoneNumber"] = this.phoneNumber;
        data["address"] = this.address;
        return data; 
    }
}

export interface IContactDataDto {
    firstName: string;
    lastName: string;
    phoneNumber: string;
    address: string;
}

export class ContactDto extends ContactDataDto implements IContactDto {
    id!: string;

    constructor(data?: IContactDto) {
        super(data);
    }

    init(_data?: any, _mappings?: any) {
        super.init(_data);
        if (_data) {
            this.id = _data["id"];
        }
    }

    static fromJS(data: any, _mappings?: any): ContactDto {
        data = typeof data === 'object' ? data : {};
        return createInstance<ContactDto>(data, _mappings, ContactDto);
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        super.toJSON(data);
        return data; 
    }
}

export interface IContactDto extends IContactDataDto {
    id: string;
}

export class PagingDto implements IPagingDto {
    offset!: number;
    pageRecords!: number;
    allRecords!: number;

    constructor(data?: IPagingDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any, _mappings?: any) {
        if (_data) {
            this.offset = _data["offset"];
            this.pageRecords = _data["pageRecords"];
            this.allRecords = _data["allRecords"];
        }
    }

    static fromJS(data: any, _mappings?: any): PagingDto {
        data = typeof data === 'object' ? data : {};
        return createInstance<PagingDto>(data, _mappings, PagingDto);
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["offset"] = this.offset;
        data["pageRecords"] = this.pageRecords;
        data["allRecords"] = this.allRecords;
        return data; 
    }
}

export interface IPagingDto {
    offset: number;
    pageRecords: number;
    allRecords: number;
}

export class ModelErrorDto implements IModelErrorDto {
    modelType!: string | undefined;
    errors!: { [key: string]: string[]; } | undefined;

    constructor(data?: IModelErrorDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any, _mappings?: any) {
        if (_data) {
            this.modelType = _data["modelType"];
            if (_data["errors"]) {
                this.errors = {} as any;
                for (let key in _data["errors"]) {
                    if (_data["errors"].hasOwnProperty(key))
                        this.errors![key] = _data["errors"][key] !== undefined ? _data["errors"][key] : [];
                }
            }
        }
    }

    static fromJS(data: any, _mappings?: any): ModelErrorDto {
        data = typeof data === 'object' ? data : {};
        return createInstance<ModelErrorDto>(data, _mappings, ModelErrorDto);
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["modelType"] = this.modelType;
        if (this.errors) {
            data["errors"] = {};
            for (let key in this.errors) {
                if (this.errors.hasOwnProperty(key))
                    data["errors"][key] = this.errors[key];
            }
        }
        return data; 
    }
}

export interface IModelErrorDto {
    modelType: string | undefined;
    errors: { [key: string]: string[]; } | undefined;
}